Missed file (#76)
* Recurse where expected when encoding GPSTIME.
Add tests for issue 44.
Make IO tests more generic.

* Open LAZ file as binary.

* Update location of test data.

* Add (deprecated) functions to support backward-compatibility.

* Checkpoint. The encoder/decoder now sits inside the compressor/decompressor.

* Working checkpoint without record_compressor.
Record compressor is still used in examples and such.

* More checkpoint. Use new classes for 1.2 compression/decompression.

* More progress and a checkpoint. Notable in that this changes the chunk handling in io.hpp.

* Minimally working.

* Seemingly working output for point14.

* Fixup LASzip VLR

* First part of decompression - in progress.

* Add gpstime code. Checkpoint.

* Working X/Y/Z/Time

* Working encoding.

* Working writer. Tests well with random data.

* A couple more GPS bugs fixed.

* All working.

* Extra bytes VLR.

* Working with extrabytes.

* Don't allocate things on the heap.

* Remove Schema.

* More cleanup.
Guard debug.

* Version and base header files.

* Cleanup.

* Checkpoint.

* Test-passing checkpoint.

* Small interface for lazperf.

* Add simple access to file streams.
Support simple chunk encoding.

* LAZ 1.4 support in a library.

* Update in prep for emscripten.

* Working WASM generation with Emscripten.

* Update version.

* Remove lazperf-specific types.
Use generic utilities.

* Remove dead file.

* Get rid of intrinsic -- slower for some reason.

* Formatting, and remote unnecessary check for 32.

* Add missed file.

* Use project source dir.

* Don't use subdirectory.

* Fixes for Linux.

* More Linux changes.

* Use memcpy instead of reinterpret_cast.

* More Linux fixes.

* Windows warnings.

* Windows warnings.

* Quiet MSVC.

* Export functions.

* Too many parens?

* More export stuff.

* Build static and shared libs.

* Fix some odd UB according the the C++ standard.

* More random nonsense.

* Move tools so things are defined.

* Install io.hpp

* Remove fast test.

* Actions.

* Actions.

* Actions.

* Actions.

* Actions

* Action

* Action

* Action

* Action

* Action

+++ cpp/lazperf/detail/field_gpstime10.cpp
@@ -149,7 +149,7 @@ const char *Gpstime10Compressor::compress(const char *buf)
             if (curr_gpstime_diff_64 == static_cast<int64_t>(curr_gpstime_diff)) {
                 // compute multiplier between current and last integer difference
                 float multi_f = (float)curr_gpstime_diff / (float)(last_gpstime_diff[last]);
-                int multi = std::round(multi_f);
+                int multi = (int)std::round(multi_f);
 
                 // compress the residual curr_gpstime_diff in dependance on the multiplier
                 if (multi == 1) {
